+++ src/components/Header/Header.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { HeaderComponent } from './Header.style'
 import person from '../Icons/Person/male.svg'
 
-import Avatar from '../../pages/UI/uikit/Avatar/Avatar.tsx';
+import  Avatar  from '../../pages/UI/uikit/Avatar';
 
 const Header = () => {
     return (
@@ -10,8 +10,8 @@ const Header = () => {
             <Avatar
                 size={42}
                 image={person} 
-                className={'test123'}
-                title={'Katya'}
+                className={'test'}
+                title={'test'}
             />
         </HeaderComponent>
     )
