+++ src/pages/UI/uikit/Avatar/Avatar.style.tsx
