+++ src/pages/UI/uikit/Avatar/index.ts
@@ -1,0 +1,3 @@
+import { Avatar } from './Avatar'
+
+export default Avatar
